[refactor]: 로그인 관련 전역 스토어 파일 분리 (#32)
